grib: fix NullPointerException for legacy NOAA RAP models
Attempting to load a [legacy] NOAA Rapid Refresh GRIB2 file fails
with a NullPointerException.

    java -jar netcdfAll-5.10.0-SNAPSHOT.jar \
      rap_130_20170720_0000_000.grb2

This file loads correctly in netcdf-java 4.6, but it has a quirk:
it contains at least one PDS that does not define a time range.

    grib_dump -O -p section_4 \
      -w section4Length=58,numberOfTimeRange=0 \
      rap_130_20170720_0000_000.grb2

6bef2fb9f5 (Fix for issue #606, version 5., 2021-02-26) corrects
time calculations for GRIB2 PDS which have incompatible units for
**time intervals** and related time quantities. The above patch
introduces a `TimeIntervalAndUnits` parser for PDS times.

If a PDS does not define *any* time intervals, the previous
version of

    Grib2Tables.getForecastTimeInterval(Grib2Record)

returns a duration of zero—i.e., no elapsed time. This is likely
the correct interpretation if no time intervals are present. After
6bef2fb9f5, this method raises a NullPointerException instead.

This is because the inner function

    Grib2Tables.getForecastTimeInterval(Grib2Pds.PdsInterval)

returns an invalid `TimeIntervalAndUnits` when it ought to return
a zero duration instead. The calling functions do not treat this
data type as fallible and may use it in conversions that
return null.

If the PDS does not contain any time intervals, return a
"zero hour" duration instead. This restores the previous behavior.

* In legacy and modern RAP models, PDS which *do* define time
  intervals define them as zero. The interpretation of the
  "quirky" PDS remains consistent with its brethren.

* This quirk is not present in more [recent] RAP model runs, so
  this bug is unlikely to affect current RAP data.

[legacy]: https://www.ncei.noaa.gov/oa/prod-model/rapid-refresh/access/historical/analysis/201707/20170720/rap_130_20170720_0000_000.grb2

[recent]: https://www.ncei.noaa.gov/oa/prod-model/rapid-refresh/access/rap-130-13km/analysis/202504/20250401/rap_130_20250401_0000_000.grb2